Add conversion support for most activations and cleanup the code
Note: Thresholds should be padded by the lower bound values of the input
range and not by the lowest threshold values (necessry for proper
conversion of LeakyReLU-like functions).

Note: Export now allows to specify the range of input values to the
model, so it is at least possible to export sqrt and log now.