Rework flow after rebasing to custom FINN feature branch
This pulls in various changes from the whole range-analysis streamlining
effort as well as some changes to the RTL simulation backend. Changes to
the simulation backend somehow broke the manual node-by-node simulation,
but the FINN built-in onse seem to be working fine.

Makes streamlining and threshold conversion configurable/optional in
preparation of the upcoming benchmarking.